+++ xcube_smos/schema.py
@@ -25,7 +25,8 @@ from xcube.util.jsonschema import JsonDateSchema
 from xcube.util.jsonschema import JsonNumberSchema
 from xcube.util.jsonschema import JsonObjectSchema
 from xcube.util.jsonschema import JsonStringSchema
-from xcube_smos.mldataset import dgg
+from xcube_smos.mldataset.newdgg import MIN_PIXEL_SIZE
+from xcube_smos.mldataset.newdgg import NUM_LEVELS
 
 STORE_PARAMS_SCHEMA = JsonObjectSchema(
     properties=dict(
@@ -61,8 +62,8 @@ OPEN_PARAMS_SCHEMA = JsonObjectSchema(
             title='Bounding box [x1,y1, x2,y2] in geographical coordinates'
         ),
         spatial_res=JsonNumberSchema(
-            enum=[(1 << level) * dgg.MIN_PIXEL_SIZE
-                  for level in range(dgg.NUM_LEVELS)],
+            enum=[(1 << level) * MIN_PIXEL_SIZE
+                  for level in range(NUM_LEVELS)],
             title='Spatial resolution in decimal degrees.',
         ),
         time_range=JsonArraySchema(
